requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
pillow==11.0.0
requests==2.32.3
simplejson==3.19.3
soupsieve==2.6
urllib3==2.3.0
